requirements.txt
# Mock Interview Preparation AI - Hugging Face

# LangChain
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1

# Hugging Face
huggingface_hub>=0.20.0
sentence-transformers>=2.2.0

# Vector store
chromadb>=0.4.0

# API framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# HTTP for Inference API (STT/TTS)
requests>=2.28.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0

# HITL Dashboard
streamlit>=1.28.0

# Testing
pytest>=7.0
